refactor: making videocontainer component composable
