+++ app/(content)/courses/[slug]/[...id]/page.tsx
@@ -12,6 +12,10 @@ import { Mdx } from "@/components/MDX";
 import { allCourses } from "contentlayer/generated";
 // UTILS
 import { getSortedCourseChapters } from "../page";
+import { ToggleProgress } from "@/components/ToggleProgress";
+import { buttonVariants } from "@/components/ui/Button";
+import { cn } from "@/utils/helpers";
+import Link from "next/link";
 
 interface CoursePageProps {
   params: {
@@ -69,11 +73,39 @@ export default async function ChapterPage({ params }: CoursePageProps) {
       className="flex flex-col justify-start items-stretch gap-10"
     >
       <VideoContainer
-        chapterId={chapter?.slugAsParams}
         videoId={chapter?.vimeo ?? undefined}
         isFree={chapter?.free ?? false}
-        prevChapterLink={showPrevious && sortedChapters[chapter?.weight - 1].slugAsParams}
-        nextChapterLink={showNext && sortedChapters[chapter?.weight + 1].slugAsParams}
+        controls={(
+          <div className="flex justify-between items-stretch">
+            <div className="flex justify-center items-center gap-2.5">
+                {showPrevious && (
+                  <Link 
+                    href={`/courses/${sortedChapters[chapter?.weight - 1].slugAsParams}}`}
+                    className={cn(buttonVariants({
+                      variant: "secondaryOutlined",
+                      size: "sm"
+                    }))}
+                  >
+                    Play Previous
+                  </Link>
+                )}
+                {showNext && (
+                  <Link 
+                    href={`/courses/${sortedChapters[chapter?.weight + 1].slugAsParams}}`}
+                    className={cn(buttonVariants({
+                      variant: "secondaryOutlined",
+                      size: "sm"
+                    }))}
+                  >
+                    Play Next
+                  </Link>
+                )}
+              </div>
+            <div className="flex justify-center items-center gap-2.5">
+              <ToggleProgress  chapterId={chapter?.slugAsParams} isFree={chapter?.free ?? false} />
+            </div>
+          </div>
+        )}
       />
       
       <Header
