+++ components/Buttons/index.tsx
@@ -1,11 +1,9 @@
 import { CheckoutButton } from "./CheckoutButton";
 import { ManageSubscriptionButton } from "./ManageSubscriptionButton";
-import { ToggleProgressButton } from "./ToggleProgressButton";
 import { ToggleAutoPlayButton } from "./ToggleAutoPlayButton";
 
 export { 
   CheckoutButton, 
   ManageSubscriptionButton, 
-  ToggleProgressButton,
   ToggleAutoPlayButton
 };
